refactor: store 하나의 파일로 변경
