--- source/frontend/src/store/login/enter.js
@@ -1,11 +1,0 @@
-const enter = {
-    state: {},
-
-    getters: {},
-
-    mutations: {},
-
-    actions:{}
-}
-
-export default enter;
