--- source/frontend/src/store/main/reports.js
@@ -1,11 +1,0 @@
-const reports = {
-    state: {},
-
-    getters: {},
-
-    mutations: {},
-
-    actions:{}
-}
-
-export default reports;
