--- source/frontend/src/store/views/slideshow.js
@@ -1,11 +1,0 @@
-const slideshow = {
-    state: {},
-
-    getters: {},
-
-    mutations: {},
-
-    actions:{}
-}
-
-export default slideshow;
